+++ twitter4j-v2-support-kotlin-example/src/main/kotlin/twitter4j_v2_support_example/App.kt
@@ -33,6 +33,10 @@ fun main(@Suppress("UNUSED_PARAMETER") args: Array<String>) {
     val conf = ConfigurationBuilder()
             .setJSONStoreEnabled(true)
             .build()
+
+    // set another baseURL of Twitter v2 API
+//    conf.v2Configuration.baseURL = "https://api.twitter.com/2/"
+
     val twitter = TwitterFactory(conf).instance
 //    val twitter = TwitterFactory.getSingleton()