+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 Change Log
 ==========
 
-v0.1.2 (2020.07.xx)
+v0.1.2 (2020.08.xx)
 -------------------
+- Add `Configuration.v2Configuration.baseURL` to set another base URL
 - Replace the method for storing json with Twitter4J style
 - Add java example
 - Rename `twitter4j-v2-support-example` to `twitter4j-v2-support-kotlin-example`